+++ bin/faucet/src/main.rs
@@ -397,8 +397,11 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
     Ok(())
 }
 
+// INTEGRATION TESTS
+// ================================================================================================
+
 #[cfg(test)]
-mod test {
+mod tests {
     use std::env::temp_dir;
     use std::process::Stdio;
     use std::str::FromStr;
@@ -414,6 +417,7 @@ mod test {
     };
     use serde_json::{Map, json};
     use tokio::io::AsyncBufReadExt;
+    use tokio::net::TcpListener;
     use tokio::time::sleep;
     use url::Url;
 
@@ -422,11 +426,12 @@ mod test {
     use crate::{Cli, run_faucet_command};
 
     /// This test starts a stub node, a faucet connected to the stub node, and a chromedriver
-    /// to test the faucet website. It then loads the website and checks that all the requests
-    /// made return status 200.
+    /// to test the faucet website. It then loads the website, mints tokens, and checks that all the
+    /// requests returned status 200.
     #[tokio::test]
-    async fn test_website() {
-        let website_url = Box::pin(start_test_faucet()).await;
+    async fn frontend_mint_tokens() {
+        let stub_node_url = run_stub_node().await;
+        let website_url = run_faucet_server(stub_node_url).await;
         let client = start_fantoccini_client().await;
 
         // Open the website
@@ -493,18 +498,22 @@ mod test {
         client.close().await.unwrap();
     }
 
-    async fn start_test_faucet() -> Url {
-        let stub_node_url = Url::from_str("http://localhost:50051").unwrap();
+    // TESTING HELPERS
+    // ---------------------------------------------------------------------------------------------
 
-        // Start the stub node
+    async fn run_stub_node() -> Url {
+        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
+        let listener_addr = listener.local_addr().unwrap();
+        let stub_node_url = Url::from_str(&format!("http://{listener_addr}")).unwrap();
         tokio::spawn({
             let stub_node_url = stub_node_url.clone();
             async move { serve_stub(&stub_node_url).await.unwrap() }
         });
+        stub_node_url
+    }
 
+    async fn run_faucet_server(stub_node_url: Url) -> String {
         let faucet_account_path = temp_dir().join("faucet.mac");
-
-        // Create faucet account
         Box::pin(run_faucet_command(Cli {
             command: crate::Command::CreateFaucetAccount {
                 output_path: faucet_account_path.clone(),
@@ -516,7 +525,9 @@ mod test {
         .await
         .unwrap();
 
-        // Start the faucet connected to the stub
+        let backend_url = "http://localhost:8000";
+        let frontend_url = "http://localhost:8080";
+
         // Use std::thread to launch faucet - avoids Send requirements
         std::thread::spawn(move || {
             // Create a new runtime for this thread
@@ -529,8 +540,8 @@ mod test {
             rt.block_on(async {
                 Box::pin(run_faucet_command(Cli {
                     command: crate::Command::Start {
-                        backend_url: Url::try_from("http://localhost:8000").unwrap(),
-                        frontend_url: Some(Url::parse("http://localhost:8080").unwrap()),
+                        backend_url: Url::try_from(backend_url).unwrap(),
+                        frontend_url: Some(Url::parse(frontend_url).unwrap()),
                         node_url: stub_node_url,
                         timeout: Duration::from_millis(5000),
                         max_claimable_amount: 1_000_000_000,
@@ -556,8 +567,8 @@ mod test {
         });
 
         // Wait for faucet to be up
-        let endpoint = Url::parse("http://localhost:8080").unwrap();
-        let addrs = endpoint.socket_addrs(|| None).unwrap();
+        let backend_url = Url::parse(backend_url).unwrap();
+        let addrs = backend_url.socket_addrs(|| None).unwrap();
         let start = Instant::now();
         let timeout = Duration::from_secs(10);
         loop {
@@ -570,7 +581,7 @@ mod test {
             }
         }
 
-        endpoint
+        frontend_url.to_string()
     }
 
     async fn start_fantoccini_client() -> fantoccini::Client {